Refactor command structure by moving and renaming command classes to shared package; remove obsolete autonomous and drive command files.
