+++ src/main/java/frc/robot/commands/shared/KSCharacterization.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.shared;
 
 import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj2.command.Command;
