+++ src/main/java/frc/robot/commands/v0_Funky/V0_FunkyCompositeCommands.java
@@ -1,0 +1,3 @@
+package frc.robot.commands.v0_Funky;
+
+public class V0_FunkyCompositeCommands {}
