+++ src/main/java/frc/robot/commands/v0_Funky/autonomous/V0_FunkyAutonomousTest.java
@@ -1,0 +1,3 @@
+package frc.robot.commands.v0_Funky.autonomous;
+
+public class V0_FunkyAutonomousTest {}
