+++ src/main/java/frc/robot/commands/v1_Gamma/V1_GammaCompositeCommands.java
@@ -1,0 +1,3 @@
+package frc.robot.commands.v1_Gamma;
+
+public class V1_GammaCompositeCommands {}
