+++ src/main/java/frc/robot/commands/v1_Gamma/autonomous/V0_GammaAutonomousTest.java
@@ -1,0 +1,3 @@
+package frc.robot.commands.v1_Gamma.autonomous;
+
+public class V0_GammaAutonomousTest {}
